pyproject.toml
[project]
name = "docker-cron-env"
version = "0.1.0"
description = "Docker環境でcronジョブを実行するサンプルプロジェクト"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "pydantic-settings==2.8.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "ruff==0.3.3",
]

[tool.ruff]
target-version = "py312"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"